requirements.txt
asttokens==2.4.1
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
comm==0.2.0
crashtest==0.4.1
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
dulwich==0.21.7
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
greenlet==3.0.1
idna==3.6
installer==0.7.0
ipykernel==6.27.1
ipython==8.18.1
jaraco.classes==3.3.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.0
matplotlib-inline==0.1.6
more-itertools==10.1.0
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyMySQL==1.1.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
tomlkit==0.12.3
tornado==6.3.3
traitlets==5.14.0
trove-classifiers==2023.11.29
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.12
wheel==0.42.0
